Revert
This reverts commit 2fe177e074a915636e5bc31ecc2f7ec4b2e8fa34.

Conflicts:
	gwtapp/src/main/webapp/client/app/main/app.js
	gwtapp/src/main/webapp/index.html
	querydsl/src/main/java/com/aplana/taxaccounting/querydsl/QueryDSLGenRunner.java
	service/src/main/java/com/aplana/sbrf/taxaccounting/script/service/DeclarationService.java
	service/src/main/java/com/aplana/sbrf/taxaccounting/script/service/ReportPeriodService.java
	service/src/main/java/com/aplana/sbrf/taxaccounting/script/service/impl/DeclarationServiceImpl.java
	service/src/main/java/com/aplana/sbrf/taxaccounting/script/service/impl/ReportPeriodServiceImpl.java
	service/src/main/java/com/aplana/sbrf/taxaccounting/service/impl/DeclarationDataServiceImpl.java
	service/src/test/java/com/aplana/sbrf/taxaccounting/script/service/DeclarationServiceTest.java

+++ dao/src/main/java/com/aplana/sbrf/taxaccounting/dao/DeclarationDataDao.java
@@ -113,7 +113,7 @@ public interface DeclarationDataDao extends PermissionDao {
     /**
      * Декларация по типу и отчетному периоду подразделения + «КПП» и «Налоговый орган» + АСНУ + GUID
      */
-    DeclarationData find(int declarationTypeId, Long departmentReportPeriodId, String kpp, String oktmo, String taxOrganCode, Long asnuId, String fileName);
+    DeclarationData find(int declarationTypeId, int departmentReportPeriodId, String kpp, String oktmo, String taxOrganCode, Long asnuId, String fileName);
 
     /**
      * Поиск деклараций по имени файла
